fixed moon search bug

+++ app/models/album.rb
@@ -22,10 +22,10 @@ class Album < ApplicationRecord
 
     def self.search_albums(search)
         @api = Api.new     
-        if @api.nil? || @api.empty?
-            return
-        else   
-            @items = @api.fetch("#{search}")
+        @items = @api.fetch("#{search}")
+        if @items.nil? || @items.empty?
+            return nil
+        else    
             @all_items = @items["collection"]["items"]
             self.set_hash(@all_items)
         end
@@ -33,21 +33,25 @@ class Album < ApplicationRecord
 
     private
     def self.set_hash(items)
-        super_user = User.admin
-        super_user.albums.delete_all
-        items.each do |i|  
-            image_hash = {}
-            image_hash["href"] = i["links"][0]["href"]
-            image_hash["title"] = i["data"][0]["title"]
-            image_hash["date"] = i["data"][0]["date_created"]
-            image_hash["center"] = i["data"][0]["center"]
-            image_hash["creator"] = i["data"][0]["secondary_creator"]
-            image_hash["description"] = i["data"][0]["description"] 
-            image_hash["nasa_id"] = i["data"][0]["nasa_id"]
-            album = Album.find_or_create_by(image_hash) 
-            album.user = super_user
-            album.save
+        if items == nil
+            return nil
+        else
+            super_user = User.admin
+            super_user.albums.delete_all
+            items.each do |i|  
+                image_hash = {}
+                image_hash["href"] = i["links"][0]["href"]
+                image_hash["title"] = i["data"][0]["title"]
+                image_hash["date"] = i["data"][0]["date_created"]
+                image_hash["center"] = i["data"][0]["center"]
+                image_hash["creator"] = i["data"][0]["secondary_creator"]
+                image_hash["description"] = i["data"][0]["description"] 
+                image_hash["nasa_id"] = i["data"][0]["nasa_id"]
+                album = Album.find_or_create_by(image_hash) 
+                if album.save
+                    album.user = super_user
+                end
+            end
         end
     end
-
 end
